bug fix, producers page, google maps included

+++ src/main/java/cooperate/app/business/producer/addProducer/AddProducerCommand.java
@@ -6,6 +6,7 @@ import cooperate.infrastructure.mediation.IHandleCommand;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Component
@@ -16,8 +17,8 @@ public class AddProducerCommand implements ICommand {
     private String name;
     private String description;
     private Date memberSince;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private boolean isActive;
     private int createUserId;
 
@@ -62,19 +63,19 @@ public class AddProducerCommand implements ICommand {
         this.memberSince = memberSince;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
