+++ src/main/java/cooperate/infrastructure/dto/producer/ProducerDto.java
@@ -3,6 +3,7 @@ package cooperate.infrastructure.dto.producer;
 
 import cooperate.infrastructure.mediation.IResponse;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.Timestamp;
 
@@ -12,8 +13,8 @@ public class ProducerDto implements IResponse {
     private String name;
     private String description;
     private java.sql.Date memberSince;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private boolean isActive;
     private int createUserId;
     private java.sql.Timestamp createDate;
@@ -52,19 +53,19 @@ public class ProducerDto implements IResponse {
         this.memberSince = memberSince;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
