+++ src/main/java/cooperate/web/viewmodels/producer/ProducerCreateModel.java
@@ -3,6 +3,7 @@ package cooperate.web.viewmodels.producer;
 import org.hibernate.validator.constraints.Length;
 import org.hibernate.validator.constraints.NotBlank;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 
 public class ProducerCreateModel {
@@ -12,8 +13,8 @@ public class ProducerCreateModel {
     @NotBlank
     private String description;
     private java.sql.Date memberSince;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private boolean isActive;
 
     public String getName() {
@@ -40,19 +41,19 @@ public class ProducerCreateModel {
         this.memberSince = memberSince;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
